python 3 scehdule

+++ heatmisercontroller/schedule_functions.py
@@ -65,7 +65,7 @@ class Scheduler(object):
             raise IndexError("Schedule length not multiple of %d"%self.valuesperentry)
         pad_item = [HOUR_UNUSED, 0, 12][0:self.valuesperentry]
 
-        return schedule + pad_item * ((self.valuesperentry * self.entriesperday - len(schedule))/self.valuesperentry)
+        return schedule + pad_item * int((self.valuesperentry * self.entriesperday - len(schedule))/self.valuesperentry)
 
     def display(self):
         """Prints schedule to stdout"""
